Django Backend/FlickRankBackend/requirements.txt
absl-py==0.9.0
asgiref==3.2.7
astor==0.8.1
astroid==2.3.3
backports.functools-lru-cache==1.6.1
backports.weakref==1.0rc1
beautifulsoup4==4.8.2
bleach==3.1.4
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cycler==0.10.0
decorator==4.4.2
Django==3.0.6
djangorestframework==3.11.0
enum34==1.1.9
Flask==1.1.2
funcsigs==1.0.2
gast==0.2.2
grpcio==1.27.2
gunicorn==19.9.0
idna==2.9
imageio==2.8.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
lxml==4.5.0
MarkupSafe==1.1.1
matplotlib==3.1.3
mccabe==0.6.1
networkx==2.4
numpy==1.18.1
opencv-contrib-python==4.2.0.32
opencv-python==4.2.0.32
Pillow==7.1.2
pylint==2.4.4
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.23.0
scipy==1.4.1
selenium==3.141.0
six==1.14.0
soupsieve==2.0
spacey==0.1.1
sqlparse==0.3.1
tifffile==2020.5.25
urllib3==1.25.8
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.11.2
